% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colar.R
\name{colar}
\alias{colar}
\title{Uma funcão que junta duas strings com um espaço " "}
\usage{
colar(x, y)
}
\arguments{
\item{x}{uma string}

\item{y}{outra sring}
}
\description{
Utiliza a função paste setando um parametro sep e aceitando duas variaveis iniciais
}
\examples{
colar()
}
\keyword{paste}
