% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colar.R
\name{colar}
\alias{colar}
\title{Cola dois numeros}
\usage{
colar(x, y)
}
\arguments{
\item{x}{primeiro argumento}

\item{y}{segundo argumento}
}
\value{
os paramentros colados
}
\description{
Recebe os parametros x e y e retorna eles colados, com um espaco de separacao
}
